#352 Pull out element fetching logic to separate file
