+++ lib/helper.js
@@ -72,6 +72,10 @@ function assert(value, message) {
         throw new Error(message);
 }
 
+const assertType = (obj, condition = isString, message = 'String parameter expected') => {
+    if (!condition(obj)) throw new Error(message);
+};
+
 const isFunction = (functionToCheck) => {
     return functionToCheck && {}.toString.call(functionToCheck) === '[object Function]';
 };
@@ -181,5 +185,6 @@ module.exports = {
     xpath,
     timeouts,
     waitUntil,
-    handleUrlRedirection
+    handleUrlRedirection,
+    assertType
 };
