+++ lib/taiko.js
@@ -3,7 +3,7 @@ const debug = require('debug');
 const logEvent = debug('taiko:event');
 const childProcess = require('child_process');
 const { helper, wait, isString, isFunction, getIfExists, waitUntil,
-    exists, xpath, waitForNavigation, timeouts } = require('./helper');
+    exists, xpath, waitForNavigation, timeouts, assertType } = require('./helper');
 const { createJsDialogEventName } = require('./util');
 const inputHandler = require('./handlers/inputHandler');
 const domHandler = require('./handlers/domHandler');
@@ -13,9 +13,9 @@ const targetHandler = require('./handlers/targetHandler');
 const runtimeHandler = require('./handlers/runtimeHandler');
 const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
+const {match,$$,$$xpath,initElementSearch,isSelector,elements,element,isElementVisible} = require('./elementSearch');
 const { handleRelativeSearch, setDOM, RelativeSearchElement } = require('./proximityElementSearch');
 const { setConfig, defaultConfig, setNavigationOptions, setClickOptions, setBrowserOptions } = require('./config');
-const { filterVisibleNodes } = require('./nodeFilters');
 const fs = require('fs');
 const path = require('path');
 const EventEmitter = require('events').EventEmitter;
@@ -69,6 +69,7 @@ const connect_to_cri = async (target) => {
                 });
                 await Promise.all([runtime.enable(), network.enable(), page.enable(), dom.enable(), overlay.enable(), security.enable()]);
                 await setDOM(domHandler);
+                initElementSearch(runtimeHandler,handleRelativeSearch,descEvent);
                 if (defaultConfig.ignoreSSLErrors) security.setIgnoreCertificateErrors({ ignore: true });
                 _client.on('disconnect', reconnect);
                 device = process.env.TAIKO_EMULATE_DEVICE;
@@ -2101,66 +2102,6 @@ function text(text, ...args) {
     return { get: getIfExists(get), exists: exists(getIfExists(get), descEvent), description: `Element with text "${text}"`, text: selectorText(get) };
 }
 
-function match(text, ...args) {
-    validate();
-    assertType(text);
-    const get = async (e = '*') => {
-        let elements;
-        let nbspChar = String.fromCharCode(160);
-        let textToTranslate = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ' + nbspChar;
-        let translateTo = 'abcdefghijklmnopqrstuvwxyz' + ' ';
-        let xpathText = `translate(normalize-space(${xpath(text)}), "${textToTranslate}", "${translateTo}")`;
-        elements = await matchExactElements(e, textToTranslate, translateTo, xpathText);
-        if (!elements || !elements.length) elements = await matchContainsElements(e, textToTranslate, translateTo, xpathText);
-        return await handleRelativeSearch(elements, args);
-    };
-    return { get: getIfExists(get), exists: exists(getIfExists(get), descEvent), description: `Element matching text "${text}"`, text: selectorText(get) };
-}
-
-async function matchExactElements(e, textToTranslate, translateTo, xpathText) {
-    let elements = [];
-    if (e === '*') elements = await matchTextNode(textToTranslate, translateTo, xpathText);
-    let valueAndTypeElements = await matchValueOrType(e, textToTranslate, translateTo, xpathText);
-    if (valueAndTypeElements.length) elements = elements.concat(valueAndTypeElements);
-    if (!elements || !elements.length) elements = await matchTextAcrossElements(e, textToTranslate, translateTo, xpathText);
-    return elements;
-}
-
-async function matchContainsElements(e, textToTranslate, translateTo, xpathText) {
-    let elements;
-    if (e === '*') elements = await containsTextNode(textToTranslate, translateTo, xpathText);
-    let valueAndTypeElements = await containsValueOrType(e, textToTranslate, translateTo, xpathText);
-    if (valueAndTypeElements.length) elements = elements.concat(valueAndTypeElements);
-    if (!elements || !elements.length) elements = await containsTextAcrossElements(e, textToTranslate, translateTo, xpathText);
-    return elements;
-}
-
-
-async function matchTextNode(textToTranslate, translateTo, xpathText) {
-    return await $$xpath(`//text()[translate(normalize-space(string()), "${textToTranslate}", "${translateTo}")=${xpathText}]`);
-}
-
-async function containsTextNode(textToTranslate, translateTo, xpathText) {
-    return await $$xpath(`//text()[contains(translate(normalize-space(string()), "${textToTranslate}", "${translateTo}"), ${xpathText})]`);
-}
-
-async function matchValueOrType(e, textToTranslate, translateTo, xpathText) {
-    return await $$xpath('//' + e + `[translate(normalize-space(@value), "${textToTranslate}", "${translateTo}")=${xpathText} or translate(normalize-space(@type), "${textToTranslate}", "${translateTo}")=${xpathText}]`);
-}
-
-async function containsValueOrType(e, textToTranslate, translateTo, xpathText) {
-    return await $$xpath('//' + e + `[contains(translate(normalize-space(@value), "${textToTranslate}", "${translateTo}"), ${xpathText}) or contains(translate(normalize-space(@type), "${textToTranslate}", "${translateTo}"), ${xpathText})]`);
-}
-
-async function matchTextAcrossElements(e, textToTranslate, translateTo, xpathText) {
-    const xpathToTranslateInnerText = `translate(normalize-space(.), "${textToTranslate}", "${translateTo}")`;
-    return await $$xpath('//' + e + `[not(descendant::*[${xpathToTranslateInnerText}=${xpathText}]) and ${xpathToTranslateInnerText}=${xpathText}]`);
-}
-
-async function containsTextAcrossElements(e, textToTranslate, translateTo, xpathText) {
-    return await $$xpath('//' + e + `[not(descendant::*[contains(translate(normalize-space(.), '${textToTranslate}', '${translateTo}'), ${xpathText})]) and contains(translate(normalize-space(.), '${textToTranslate}', '${translateTo}'), ${xpathText})]`);
-}
-
 /**
  * Search relative HTML elements with this {@link relativeSelector}.
  *
@@ -2672,31 +2613,12 @@ const highlightElemOnAction = async (nodeId) => {
             await waitFor(1000);
             await overlay.hideHighlight();
         } catch(err) {
-            if((await filterVisibleNodes([nodeId], evaluate)).length < 1 )
+            if(await isElementVisible(nodeId))
                 console.warn('WARNING: Taiko cannot highlight hidden elements.');
             else
                 throw err;
         }
     }
-};
-
-const element = async (selector, tag) => (await elements(selector, tag))[0];
-
-const elements = async (selector, tag) => {
-    const elements = await (() => {
-        if (isString(selector)) {
-            return match(selector).get(tag);
-        } else if (isSelector(selector)) {
-            return selector.get(tag);
-        }
-        return null;
-    })();
-    if (!elements || !elements.length) {
-        const error = isString(selector) ? `Element with text ${selector} not found` :
-            `${selector.description} not found`;
-        throw new Error(error);
-    }
-    return elements;
 };
 
 const description = (selector, lowerCase = false) => {
@@ -2750,26 +2672,6 @@ const dialog = (dialogType, dialogMessage, callback) => {
     });
 };
 
-const isSelector = obj => obj && obj['get'] && obj['exists'];
-
-const $$ = async (selector) => {
-    const elements = await runtimeHandler.findElements(`document.querySelectorAll('${selector}')`);
-    return (await filterVisibleNodes(elements, evaluate));
-};
-
-const $$xpath = async (selector, selectHiddenElements) => {
-    var xpathFunc = function (selector) {
-        var result = [];
-        var nodesSnapshot = document.evaluate(selector, document, null, XPathResult.ORDERED_NODE_SNAPSHOT_TYPE, null);
-        for (var i = 0; i < nodesSnapshot.snapshotLength; i++) {
-            result.push(nodesSnapshot.snapshotItem(i));
-        }
-        return result;
-    };
-    const elements = await runtimeHandler.findElements(xpathFunc, selector);
-    return selectHiddenElements ? elements : (await filterVisibleNodes(elements, evaluate));
-};
-
 const evaluate = async (selector, func) => {
     let nodeId = selector;
     if (isNaN(selector)) nodeId = await element(selector);
@@ -2780,11 +2682,6 @@ const evaluate = async (selector, func) => {
 const validate = () => {
     if (!dom || !page) throw new Error('Browser or page not initialized. Call `openBrowser()` before using this API');
 };
-
-const assertType = (obj, condition = isString, message = 'String parameter expected') => {
-    if (!condition(obj)) throw new Error(message);
-};
-
 
 const selectorText = get => {
     return async () => {
