+++ test/unit-tests/elementSearch.test.js
@@ -1,15 +1,16 @@
 const expect = require('chai').expect;
 const rewire = require('rewire');
-const nodeFilters = rewire('../../lib/nodeFilters');
+const elementSearch = rewire('../../lib/elementSearch');
 class DomRects {
 }
 const TEXT_NODE = 3;
-describe('Node filters', () => {
+describe('Element Search', () => {
 
     describe('Filter visible nodes', () => {
         let nodeIds;
+        const filterVisibleNodes = elementSearch.__get__('filterVisibleNodes');
         beforeEach(() => {
-            nodeFilters.__set__('Node', { TEXT_NODE });
+            elementSearch.__set__('Node', { TEXT_NODE });
             nodeIds = {
                 23: {
                     offsetHeight: 1, offsetWidth: 0
@@ -29,19 +30,19 @@ describe('Node filters', () => {
                     }
                 }
             };
+            elementSearch.__set__('runtimeHandler',{runtimeCallFunctionOn:(predicate,args,obj)=>{
+                const result = {result:{value:predicate.call(nodeIds[obj.nodeId])}};
+                return result;
+            }});
         });
         it('should return visible nodes', async () => {
-            const callOnFunction = (nodeId, predicate) => ({ value: predicate.call(nodeIds[nodeId]) });
             const visibleNodeIds = [23, 45, 67];
-            expect(await nodeFilters.filterVisibleNodes([23, 45, 67, 68], callOnFunction)).to.eql(visibleNodeIds);
+            expect(await filterVisibleNodes([23, 45, 67, 68])).to.eql(visibleNodeIds);
         });
 
         it('should use parentElement to determine visibility for text nodes', async () => {
-            let callOnFunction = (nodeId, predicate) => {
-                return { value: predicate.call(nodeIds[nodeId]) };
-            };
             const visibleNodeIds = [23, 89];
-            expect(await nodeFilters.filterVisibleNodes([23, 68, 89], callOnFunction)).to.eql(visibleNodeIds);
+            expect(await filterVisibleNodes([23, 68, 89])).to.eql(visibleNodeIds);
         });
     });
 });
